Refactor environment config and add dev/test support
Moved environment.ts to environments/ and created separate environment files for development and tests. Updated Angular build configuration to use file replacements for different environments. Refactored imports across services and components to use the new environment path and switched OAuth config to use environment variables for URLs and client IDs.
