+++ frontend/angular.json
@@ -51,7 +51,34 @@
             "development": {
               "optimization": false,
               "extractLicenses": false,
-              "sourceMap": true
+              "sourceMap": true,
+              "fileReplacements": [
+                {
+                  "replace": "src/environments/environment.ts",
+                  "with": "src/environments/environment.development.ts"
+                }
+              ]
+            },
+            "tests": {
+              "budgets": [
+                {
+                  "type": "initial",
+                  "maximumWarning": "2MB",
+                  "maximumError": "5MB"
+                },
+                {
+                  "type": "anyComponentStyle",
+                  "maximumWarning": "4kB",
+                  "maximumError": "8kB"
+                }
+              ],
+              "outputHashing": "all",
+              "fileReplacements": [
+                {
+                  "replace": "src/environments/environment.ts",
+                  "with": "src/environments/environment.tests.ts"
+                }
+              ]
             }
           },
           "defaultConfiguration": "production"
@@ -64,6 +91,9 @@
             },
             "development": {
               "buildTarget": "frontend:build:development"
+            },
+            "tests": {
+              "buildTarget": "frontend:build:tests"
             }
           },
           "defaultConfiguration": "development"
