+++ frontend/src/app/services/statistics.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { environment } from '../../environment';
+import { environment } from '../../environments/environment';
 import { Statistics } from '../models/statistics';
 
 @Injectable({
