+++ frontend/src/environments/environment.ts
@@ -1,6 +1,5 @@
 export const environment = {
     // This is the environment configuration for the frontend application.
-    // apiUrl: 'http://localhost:8080',
     apiUrl: 'https://api.insy.hs-esslingen.com',
 
     // Regex to match price format with optional Euro sign
@@ -8,4 +7,7 @@ export const environment = {
     // A euro sign is optional and can be separated by a space
     // Whitespace is allowed before and after the price
     priceRegEx: /^\s*\d+(?:[.,]\d{1,2})?\s*€?\s*$/,
+
+    clientId: "angular-app-prod",
+    issuer: "https://auth.insy.hs-esslingen.com/realms/insy",
 }
